emma_service/requirements.txt
# Emma Service Dependencies

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# OpenAI for Emma AI
openai>=1.12.0

# HTTP requests
requests>=2.31.0
httpx>=0.26.0

# Database (for dashboard_bridge)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Environment
python-dotenv>=1.0.0
